# Fix file search performance regression