@ewels 2024-08-13
Great stuff, thanks for noticing and fixing @rhpvorderman!

> Could you please ping me next time there is a big performance regression? Usually these are relatively easy to fix.

I guess this went unnoticed. @vladsavelyev any ideas how we could flag regressions like this in CI? 🤔  Nothing really springs to mind for me, aside from hardcoding some expected values to check for somehow.