@rhpvorderman 2024-08-14
Testing performance regressions is really hard, I don't know how to tackle that other than to manually run things. I am very happy with the unit tests you added though. I feel more confident now that I have not broken anything again for the sake of performance :sweat_smile: .